ocl: move deprecated code from ocl.cpp
